requirements.txt
numpy
pycryptodome
pillow
matplotlib
prettytable
flask